project/python/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
supabase==2.3.4
httpx==0.25.2
pandas==2.1.4
numpy==1.24.3
requests==2.31.0
python-multipart==0.0.6